requirements.txt
numpy>=1.20.3
opencv-python==4.5.5.64
pandas==1.3.4
Pillow==9.4.0
PyYAML==6.0
scipy==1.7.1
six==1.16.0
tqdm==4.62.3
fvcore